Add ItemPolicy, Middleware, ImageUpload

+++ app/Http/Controllers/ItemController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Item;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ItemController extends Controller
 {
@@ -14,8 +16,15 @@ class ItemController extends Controller
     public function index()
     {
         $items = Item::query()
+
             ->orderBy('name')
             ->get();
+
+        // if (Auth::user()->role == 'admin') {
+        //     $items = Item::query()
+        //         ->orderBy('name')
+        //         ->get();
+        // }
 
         return view('items.index', [
             'items' => $items
@@ -27,6 +36,8 @@ class ItemController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Item::class);
+
         return view('items.create');
     }
 
@@ -35,6 +46,8 @@ class ItemController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Item::class);
+
         // validation
         $validated = $request->validate([
             'name' => ['required', 'string', 'max:255'],
@@ -47,10 +60,22 @@ class ItemController extends Controller
         // $item->name = $validated['name'];
         // $item->sku = $validated['sku'];
         // $item->reorder_level = $validated['reorder_level'];
+        // $item->user_id = Auth::id();
         // $item->save();
 
         // fancy way
-        Item::create($validated);
+        // Item::create([
+        //     'name' => $validated['name'],
+        //     'sku' => $validated['sku'],
+        //     'reorder_level' => $validated['reorder_level'],
+        //     'user_id' => Auth::id(),
+        // ]);
+
+        // another fancy way
+        Item::create(['user_id' => Auth::id(), ...$validated]);
+
+        // easy way
+        // Auth::user()->items()->create($validated);
 
         return redirect('/items')->with('success', 'Item created');
     }
@@ -60,6 +85,8 @@ class ItemController extends Controller
      */
     public function show(Item $item)
     {
+        Gate::authorize('view', $item);
+
         return view('items.show', [
             'item' => $item
         ]);
@@ -70,6 +97,8 @@ class ItemController extends Controller
      */
     public function edit(Item $item)
     {
+        Gate::authorize('update', $item);
+
         return view('items.edit', [
             'item' => $item
         ]);
@@ -80,6 +109,8 @@ class ItemController extends Controller
      */
     public function update(Request $request, Item $item)
     {
+        Gate::authorize('update', $item);
+
         //validation
         $validated = $request->validate([
             'name' => ['required', 'string', 'max:255'],
@@ -104,6 +135,8 @@ class ItemController extends Controller
      */
     public function destroy(Item $item)
     {
+        Gate::authorize('delete', $item);
+
         $item->delete();
 
         return redirect('/items')->with('success', 'Item deleted');
